infra/scripts/requirements.txt
azure-storage-blob
azure-storage-queue
aiohttp